style: restructure location related files
